split paletted_image to its own file, create constants file
