+++ src/main.cpp
@@ -19,37 +19,12 @@
 #define JO_GIF_HEADER_FILE_ONLY
 #include <jo_gif.cpp>
 
+#include "constants.hpp"
 #include "arrays.hpp"
 #include "rgba.hpp"
 #include "options.hpp"
 #include "result.hpp"
-
-const auto kUsage = R"(
-wfc.bin [-h/--help] [--gif] [job=samples.cfg, ...]
-	-h/--help   Print this help
-	--gif       Export GIF images of the process
-	file        Jobs to run
-)";
-
-using emilib::irange;
-
-using Bool              = uint8_t; // To avoid problems with vector<bool>
-using ColorIndex        = uint8_t; // tile index or color index. If you have more than 255, don't.
-using Palette           = std::vector<RGBA>;
-using Pattern           = std::vector<ColorIndex>;
-using PatternHash       = uint64_t; // Another representation of a Pattern.
-using PatternPrevalence = std::unordered_map<PatternHash, size_t>;
-using RandomDouble      = std::function<double()>;
-using PatternIndex      = uint16_t;
-
-const auto kInvalidIndex = static_cast<size_t>(-1);
-const auto kInvalidHash = static_cast<PatternHash>(-1);
-
-const bool   kGifSeparatePalette  = true;
-const size_t kGifInterval         =  16; // Save an image every X iterations
-const int    kGifDelayCentiSec    =   1;
-const int    kGifEndPauseCentiSec = 200;
-const size_t kUpscale             =   4; // Upscale images before saving
+#include "paletted_image.hpp"
 
 
 const char* result2str(const Result result)
@@ -63,17 +38,6 @@ const size_t MAX_COLORS = 1 << (sizeof(ColorIndex) * 8);
 
 using Graphics = Array2D<std::vector<ColorIndex>>;
 
-struct PalettedImage
-{
-	size_t                  width, height;
-	std::vector<ColorIndex> data; // width * height
-	Palette                 palette;
-
-	ColorIndex at_wrapped(size_t x, size_t y) const
-	{
-		return data[width * (y % height) + (x % width)];
-	}
-};
 
 // What actually changes
 struct Output
